% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_refs.R
\docType{data}
\name{code_refs}
\alias{code_refs}
\title{code_refs}
\format{
A data frame with 1 rows and 3 variables:
\describe{
\item{country_code}{Country code (iso2)}
\item{url_ref}{Url for specific RGS dataset}
\item{args_read}{Arguments for reading the data file}
}
}
\source{
various sources
}
\usage{
code_refs
}
\description{
A dataset for finding the country specific RGS system.
}
\keyword{datasets}
