% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox.R
\name{evaluate_model_cox}
\alias{evaluate_model_cox}
\title{Evaluate model with cutoffs
Evaluates the model by creating exposure variables based on the cutoffs and fitting a Cox model.}
\usage{
evaluate_model_cox(my_data, cutoffs, formula)
}
\arguments{
\item{my_data}{A dataframe containing the data.}

\item{cutoffs}{A list of dataframes containing the cutoffs for each variable.}

\item{formula}{An object of class "formula" describing the model structure.}
}
\value{
A dataframe summarizing the model evaluation metrics.
}
\description{
Evaluate model with cutoffs
Evaluates the model by creating exposure variables based on the cutoffs and fitting a Cox model.
}
