% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{evaluate_model_logistic}
\alias{evaluate_model_logistic}
\title{Evaluate logistic regression model with cutoffs}
\usage{
evaluate_model_logistic(my_data, cutoffs, formula)
}
\arguments{
\item{my_data}{A dataframe containing the data.}

\item{cutoffs}{A list of dataframes containing the cutoffs for each variable.}

\item{formula}{An object of class "formula" describing the model structure.}
}
\value{
A dataframe summarizing the model evaluation metrics.
}
\description{
This function evaluates the logistic regression model by creating exposure variables based on the cutoffs and fitting the model.
}
