% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\docType{data}
\name{examples}
\alias{examples}
\title{examples}
\format{
A data frame with 128 rows and 2 variables:
\describe{
\item{daybook}{Journal post}
\item{fixed}{Reference codes associated with the journal post}
}
}
\source{
various sources
}
\usage{
examples
}
\description{
A dataset for specific ledger examples.
}
\keyword{datasets}
