% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\docType{data}
\name{fdata}
\alias{fdata}
\title{Data to show something}
\format{
data frame with 100 rows and 3 variables

\describe{
    \item{the first col}{Y}
    \item{the second}{X1}
    \item{the third col}{X2}
    }
}
\source{
{create the dataset}
}
\usage{
fdata
}
\description{
Contains the important data
}
\examples{
data(fdata)

}
\keyword{datasets}
