% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{find_best_combination_logistic}
\alias{find_best_combination_logistic}
\title{Find best combination of cutoffs for logistic regression}
\usage{
find_best_combination_logistic(my_data, cutoffs_list, formula)
}
\arguments{
\item{my_data}{A dataframe containing the data.}

\item{cutoffs_list}{A list of dataframes containing the cutoffs for each variable.}

\item{formula}{An object of class "formula" describing the model structure.}
}
\value{
A dataframe containing the optimal cutoff values and the AIC value.
}
\description{
This function searches for the best combination of cutoffs that minimizes the Akaike Information Criterion (AIC).
}
