% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox.R
\name{find_cutoffs_cox}
\alias{find_cutoffs_cox}
\title{Find symmetric cutoffs
This function finds symmetric cutoff points for a specified variable using Cox regression.}
\usage{
find_cutoffs_cox(my_data, variable_name)
}
\arguments{
\item{my_data}{A dataframe containing the data.}

\item{variable_name}{A character string specifying the name of the variable for which to find cutoffs.}
}
\value{
A dataframe containing the symmetric cutoffs and their corresponding predicted log hazard ratios.
}
\description{
Find symmetric cutoffs
This function finds symmetric cutoff points for a specified variable using Cox regression.
}
