% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{find_cutoffs_logistic}
\alias{find_cutoffs_logistic}
\title{Find symmetric cutoffs for logistic regression}
\usage{
find_cutoffs_logistic(my_data, variable)
}
\arguments{
\item{my_data}{A dataframe containing the data.}

\item{variable}{A character string specifying the name of the predictor variable for which to find cutoffs.}
}
\value{
A dataframe containing the symmetric cutoffs and their corresponding predicted log odds ratios.
}
\description{
This function finds symmetric cutoff points for a specified predictor variable using logistic regression.
}
