% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference-codes.R
\name{get_standard_business_reporting}
\alias{get_standard_business_reporting}
\title{Obtain country specific reference system.}
\usage{
get_standard_business_reporting(country = "nl")
}
\arguments{
\item{country}{A character string of the countries names or abbreviation.}
}
\value{
One or more entries form the coder reference database in tibble
format.
}
\description{
Often accountants use standardized reference codes to assign booking to
certain journal posts and account on the general ledger.
}
