% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_module.R
\name{input_ui}
\alias{input_ui}
\alias{output_ui}
\alias{input_server}
\alias{output_server}
\title{Dataset selection}
\usage{
input_ui(id, filter = NULL)

output_ui(id)

input_server(id)

output_server(id, augmented)
}
\description{
Dataset selection
}
