% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_S.R
\name{plot_S}
\alias{plot_S}
\title{Display of Survival Curves per Study Design}
\usage{
plot_S(
  S = list(S0, S1),
  Tmax = 50,
  leg = list(x = 30, y = 1, txt = c("Control", "Exp. Arm")),
  param = list(xlab = "Survival Time (mo)", ylab = "PFS", main =
    "PFS Survival Curve Per Study Design")
)
}
\arguments{
\item{S}{An object containly a list of survival functions to draw}

\item{t}{A sequence of survival time for the plot (x-axis)}

\item{...}{Other graphic parameters passed to the plot}
}
\value{
Display of the graph
}
\description{
This function plots the survival curves per study design
}
\examples{

HR = 0.65; delay = 6; lambda0 = log(2) / 12; 
h0 = function(t){lambda0}; S0 = function(t){exp(-lambda0 * t)}
Hazard function and survival function for experimental arm
lambda1 = lambda0 * HR
h1 = function(t){lambda1}; S1= function(t){exp(-lambda1 * t)}
F.entry = function(t){(t/18)^1.5*as.numeric(t <= 18) + as.numeric(t > 18)}

plotS(Tmax = 50, S = list(S0, S1), param=list(xlab="Time (mo)", ylab="PFS", main="PFS: HR = 0.65; Design Assumptions"))

}
