% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGS_app.R
\name{RGS_app}
\alias{RGS_app}
\title{Explore RGS data app}
\usage{
RGS_app()
}
\value{
Shiny app.
}
\description{
Shiny app for exploring of RGS
(\href{https://www.referentiegrootboekschema.nl/}{Referentie GrootboekSchema})
datasets.
}
