% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-select_module.R
\name{select_ui}
\alias{select_ui}
\alias{select_server}
\title{Variable selection}
\usage{
select_ui(id)

select_server(
  id,
  RGS,
  default = c("referentiecode", "omschrijving", "referentienummer")
)
}
\description{
Variable selection
}
