% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_module.R
\name{table_ui}
\alias{table_ui}
\alias{table_server}
\title{Hierarchical table shiny module}
\usage{
table_ui(id, select, download, level = "Niveau", ref_code = "Referentiecode")

table_server(id, RGS, labels = "Niveau ")
}
\arguments{
\item{id}{Namespace id.}

\item{select}{Add variable selection interface.}

\item{download}{Add download button.}

\item{level}{Name on controller ui for level selection.}

\item{ref_code}{Name on controller ui for reference code selection.}

\item{RGS}{Reactive value of the RGS data.}

\item{labels}{Labels for hierarchical structure (default = \code{"Niveau "}.}
}
\value{
Shiny GUI or server
}
\description{
This module creates the table input and output as an
\code{reactable::\link[reactable:reactable]{reactable}()}.
}
