% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-filter_module.R
\name{filter_ui}
\alias{filter_ui}
\alias{filter_server}
\title{Observation filtering}
\usage{
filter_ui(id)

filter_server(
  id,
  RGS,
  external,
  iexternal = "referentiecode",
  ivars = c(Balanszijde = "d_c", Niveau = "nivo")
)
}
\arguments{
\item{id}{Namespace id.}

\item{RGS}{Reactive value for the reference system.}

\item{external}{Reactive value for children of parent RGS.}
}
\value{
Shiny GUI or server.
}
\description{
GUI and server modules for filtering RGS datasets.
}
