% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_module.R, R/table_module.R
\name{plot_ui}
\alias{plot_ui}
\alias{plot_server}
\alias{table_server}
\title{Hierarchical plot shiny module}
\usage{
plot_ui(id, output = c("plot", "description"), download = TRUE)

plot_server(id, RGS, child)

table_server(id, RGS, labels = "Niveau ")
}
\arguments{
\item{id}{Namespace id}

\item{output}{Plot output}
}
\value{
Shiny GUI or server
}
\description{
Hierarchical plot shiny module
}
