% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_module.R
\name{plot_ui}
\alias{plot_ui}
\alias{plot_server}
\title{Hierarchical plot shiny module}
\usage{
plot_ui(id, output = c("plot", "description"), download = TRUE)

plot_server(id, RGS)
}
\arguments{
\item{id}{Namespace id.}

\item{output}{Plot output.}

\item{download}{Add download button.}

\item{RGS}{Reactive value for the reference system.}
}
\value{
Shiny GUI or server.
}
\description{
GUI and server modules for hierarchical plotting of RGS datasets.
}
