% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunburst.R
\name{RGS_sunburst}
\alias{RGS_sunburst}
\alias{parent_seeker}
\title{Sunburst}
\usage{
RGS_sunburst(
  RGS = parent_seeker(get_standard_business_reporting("Nederland")),
  interactive = TRUE
)

parent_seeker(RGS = get_standard_business_reporting("Nederland"))
}
\arguments{
\item{RGS}{\href{https://www.referentiegrootboekschema.nl/}{Referentie GrootboekSchema}}

\item{interactive}{Make plot interactive.}
}
\value{
\code{ggplot2::\link[ggplot2:ggplot]{ggplot}()}
}
\description{
Sunburst plot for hierarchical data.
}
