% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-select_module.R
\name{select_ui}
\alias{select_ui}
\alias{select_server}
\title{Variable selection}
\usage{
select_ui(id)

select_server(
  id,
  RGS,
  ref_code = c("referentiecode", "nivo", "terminal"),
  default = c("omschrijving", "referentienummer"),
  additional = c("sortering", "omschrijving_verkort", "referentie_omslagcode")
)
}
\arguments{
\item{id}{Namespace id.}

\item{RGS}{Reactive value for the reference system.}

\item{ref_code}{Pre-selected variables (character string or vector).}

\item{default}{Default variables (character string or vector).}

\item{additional}{Additional variables (character string or vector).}
}
\value{
Shiny GUI or server.
}
\description{
GUI and server modules for variable selection of RGS datasets.
}
