% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_module.R
\name{table_ui}
\alias{table_ui}
\title{Hierarchical table shiny module}
\usage{
table_ui(id, select, download = TRUE)
}
\arguments{
\item{id}{Namespace id.}

\item{RGS}{Reactive value of the RGS data.}

\item{labels}{Labels for hierarchical structure (default = \code{"Niveau "}.}
}
\value{
Shiny GUI or server
}
\description{
This module creates the table input and output.
}
